Update `socket-tls/ca/gen-certs.sh` to generate valid server cert.

+++ socket-tls/ca/gen-certs.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+cd $(dirname $(realpath $0))
+if [ ! -f $(basema,e $0) ]; then
+  exit 1
+fi
+
 rm ca-* server-* client-*
 
 echo "Generate the self-signed Root CA certificate:"
@@ -16,11 +21,20 @@ openssl genrsa -out server-key.pem 4096
 echo "Generate the server signing request:"
 
 openssl req -new -key server-key.pem -out server-csr.pem \
-  -subj "/C=DK/ST=Denmark/L=Copenhagen/O=Trashers Test Server/OU=Test Server/CN=server.js/emailAddress=server-admin@example.com"
+  -subj "/C=DK/ST=Denmark/L=Copenhagen/O=Trashers Test Server/OU=Test Server/CN=server.js/emailAddress=server-admin@example.com" \
+  || {
+    echo "Generate the server signing request: FAILED"
+    exit 1
+  }
 
 echo "Signing the server signing request:"
 
-openssl x509 -req -days 365 -in server-csr.pem -CA ca-crt.pem -CAkey ca-key.pem -CAcreateserial -out server-crt.pem -passin pass:password
+openssl x509 -req -days 365 -in server-csr.pem -CA ca-crt.pem -CAkey ca-key.pem -CAcreateserial -out server-crt.pem -passin pass:password \
+  -extfile <(printf "subjectAltName=DNS:localhost,DNS:127.0.0.1,IP:127.0.0.1") \
+  || {
+    echo "Generate the server signing: FAILED"
+    exit 1
+  }
 
 echo "Generate the client private key:"
